move safe test cases from TestQueryNode to TestContainerOperations
