+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/hdds/scm/node/TestQueryNode.java
@@ -22,26 +22,19 @@ import static org.apache.hadoop.hdds.HddsConfigKeys.HDDS_COMMAND_STATUS_REPORT_I
 import static org.apache.hadoop.hdds.HddsConfigKeys.HDDS_CONTAINER_REPORT_INTERVAL;
 import static org.apache.hadoop.hdds.HddsConfigKeys.HDDS_NODE_REPORT_INTERVAL;
 import static org.apache.hadoop.hdds.HddsConfigKeys.HDDS_PIPELINE_REPORT_INTERVAL;
-import static org.apache.hadoop.hdds.protocol.proto.HddsProtos.NodeOperationalState.DECOMMISSIONING;
-import static org.apache.hadoop.hdds.protocol.proto.HddsProtos.NodeOperationalState.IN_MAINTENANCE;
-import static org.apache.hadoop.hdds.protocol.proto.HddsProtos.NodeOperationalState.IN_SERVICE;
 import static org.apache.hadoop.hdds.protocol.proto.HddsProtos.NodeState.DEAD;
-import static org.apache.hadoop.hdds.protocol.proto.HddsProtos.NodeState.HEALTHY;
 import static org.apache.hadoop.hdds.protocol.proto.HddsProtos.NodeState.STALE;
 import static org.apache.hadoop.hdds.scm.ScmConfigKeys.OZONE_SCM_DEADNODE_INTERVAL;
 import static org.apache.hadoop.hdds.scm.ScmConfigKeys.OZONE_SCM_STALENODE_INTERVAL;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.IOException;
-import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
-import org.apache.hadoop.hdds.protocol.DatanodeDetails;
 import org.apache.hadoop.hdds.protocol.proto.HddsProtos;
 import org.apache.hadoop.hdds.scm.ScmConfigKeys;
 import org.apache.hadoop.hdds.scm.cli.ContainerOperationClient;
-import org.apache.hadoop.hdds.scm.server.StorageContainerManager;
 import org.apache.hadoop.ozone.MiniOzoneCluster;
 import org.apache.ozone.test.GenericTestUtils;
 import org.junit.jupiter.api.AfterEach;
@@ -84,13 +77,6 @@ public class TestQueryNode {
   }
 
   @Test
-  public void testHealthyNodesCount() throws Exception {
-    List<HddsProtos.Node> nodes = scmClient.queryNode(null, HEALTHY,
-        HddsProtos.QueryScope.CLUSTER, "");
-    assertEquals(numOfDatanodes, nodes.size(), "Expected live nodes");
-  }
-
-  @Test
   public void testStaleNodesCount() throws Exception {
     ExecutorService executor = Executors.newFixedThreadPool(1);
     executor.execute(() -> {
@@ -122,55 +108,5 @@ public class TestQueryNode {
     nodeCount = scmClient.queryNode(null, DEAD,
         HddsProtos.QueryScope.CLUSTER, "").size();
     assertEquals(2, nodeCount, "Mismatch of expected nodes count");
-  }
-
-  @Test
-  public void testNodeOperationalStates() throws Exception {
-    StorageContainerManager scm = cluster.getStorageContainerManager();
-    NodeManager nm = scm.getScmNodeManager();
-
-    // Set one node to be something other than IN_SERVICE
-    DatanodeDetails node = nm.getAllNodes().get(0);
-    nm.setNodeOperationalState(node, DECOMMISSIONING);
-
-    // All nodes should be returned as they are all in service
-    int nodeCount = scmClient.queryNode(IN_SERVICE, HEALTHY,
-        HddsProtos.QueryScope.CLUSTER, "").size();
-    assertEquals(numOfDatanodes - 1, nodeCount);
-
-    // null acts as wildcard for opState
-    nodeCount = scmClient.queryNode(null, HEALTHY,
-        HddsProtos.QueryScope.CLUSTER, "").size();
-    assertEquals(numOfDatanodes, nodeCount);
-
-    // null acts as wildcard for nodeState
-    nodeCount = scmClient.queryNode(IN_SERVICE, null,
-        HddsProtos.QueryScope.CLUSTER, "").size();
-    assertEquals(numOfDatanodes - 1, nodeCount);
-
-    // Both null - should return all nodes
-    nodeCount = scmClient.queryNode(null, null,
-        HddsProtos.QueryScope.CLUSTER, "").size();
-    assertEquals(numOfDatanodes, nodeCount);
-
-    // No node should be returned
-    nodeCount = scmClient.queryNode(IN_MAINTENANCE, HEALTHY,
-        HddsProtos.QueryScope.CLUSTER, "").size();
-    assertEquals(0, nodeCount);
-
-    // Test all operational states by looping over them all and setting the
-    // state manually.
-    node = nm.getAllNodes().get(0);
-    for (HddsProtos.NodeOperationalState s :
-        HddsProtos.NodeOperationalState.values()) {
-      nm.setNodeOperationalState(node, s);
-      nodeCount = scmClient.queryNode(s, HEALTHY,
-          HddsProtos.QueryScope.CLUSTER, "").size();
-      if (s == IN_SERVICE) {
-        assertEquals(5, nodeCount);
-      } else {
-        assertEquals(1, nodeCount);
-      }
-    }
   }
 }
